feat: (needs to be refactored) split files to put menu item
