+++ BE/src/main/java/com/codesquad/sidedish10/getter/response/ApiResponseForItem.java
@@ -1,0 +1,31 @@
+package com.codesquad.sidedish10.getter.response;
+
+public class ApiResponseForItem {
+  private int statusCode;
+  private int menuId;
+  private Object body;
+
+  public int getStatusCode() {
+    return statusCode;
+  }
+
+  public void setStatusCode(int statusCode) {
+    this.statusCode = statusCode;
+  }
+
+  public Object getBody() {
+    return body;
+  }
+
+  public void setBody(Object body) {
+    this.body = body;
+  }
+
+  public int getMenuId() {
+    return menuId;
+  }
+
+  public void setMenuId(int menuId) {
+    this.menuId = menuId;
+  }
+}
